+++ src/constants.ts
@@ -1,0 +1,7 @@
+export const MARKDOWN_HEADER = `翻译 \`$word\` :  
+`;
+export const MARKDOWN_FOOTER = `  
+`;
+export const MARKDOWN_LINE = `  
+*****
+`;