+++ src/index.ts
@@ -1,43 +1,9 @@
 import * as vscode from "vscode";
-import { cleanWord, getWordArray } from "./format";
-import { query } from "./query";
-
-const MARKDOWN_HEADER = `翻译 \`$word\` :  
-`;
-const MARKDOWN_FOOTER = `  
-`;
-const MARKDOWN_LINE = `  
-*****
-`;
+import { MARKDOWN_FOOTER, MARKDOWN_HEADER } from "./constants";
+import { convertToMarkdown } from "./utils/convert";
 
 // 全局翻译开关状态
 let translationEnabled = true;
-
-/**
- * 生成 Markdown 格式的翻译结果
- * @param word 单词
- * @param translation 翻译内容
- * @param phonetic 音标
- * @returns Markdown 格式的字符串
- */
-function genMarkdown(
-  word: string,
-  translation: string,
-  phonetic: string
-): string {
-  if (!translation && !phonetic) {
-    return `- [${word}](https://translate.google.com?text=${word}) :  
-本地词库暂无结果 , 查看 [Google翻译](https://translate.google.com?text=${word}) [百度翻译](https://fanyi.baidu.com/#en/zh/${word})`;
-  }
-
-  const phoneticText = phonetic ? `*/${phonetic}/*` : "";
-  return `- [${word}](https://translate.google.com?text=${word}) ${phoneticText}:  
-${translation.replace(
-  /\\n/g,
-  `  
-`
-)}`;
-}
 
 /**
  * 初始化翻译插件
@@ -120,25 +86,11 @@ export function init(context?: vscode.ExtensionContext): void {
       if (selectText && word.includes(selectText)) {
         word = selectText;
       }
-
-      const originText = cleanWord(word);
-      const words = getWordArray(cleanWord(word));
-      let hoverText = "";
-
-      for (let i = 0; i < words.length; i++) {
-        const currentWord = words[i];
-        const result = query(currentWord);
-
-        if (i === 0) {
-          hoverText += genMarkdown(currentWord, result.w, result.p);
-        } else {
-          hoverText +=
-            MARKDOWN_LINE + genMarkdown(currentWord, result.w, result.p);
-        }
-      }
+      const originText = word.replace(/"/g, "");
+      const wordsMarkdown = convertToMarkdown(word);
 
       const header = MARKDOWN_HEADER.replace("$word", originText);
-      hoverText = header + hoverText + MARKDOWN_FOOTER;
+      const hoverText = header + wordsMarkdown + MARKDOWN_FOOTER;
 
       return new vscode.Hover(hoverText);
     },